limbo: remove rollback check from ack
The check was needed when txn_limbo_ack() was blocking and it
would write the CONFIRM right away unless the rollback mode is
enabled.

But CONFIRM writing is now async and is done in the worker fiber
which also checks for the rollback flag. No need to check it twice
then.

The worst what happens now is that the ack_count might get bumped
but it won't make difference if CONFIRM later won't be written.

This commit attempts to simplify the limbo before reworking it in
scope of #8095.

NO_DOC=refactoring
NO_CHANGELOG=refactoring
NO_TEST=tested

(cherry picked from commit 18029b4bc545ff8147ac3bf111dd75d1a9618740)