limbo: ensure higher concurrency protection
The limbo has a latch which was intended to linearize any sort of
state-changing actions like CONFIRM, PROMOTE, ROLLBACK, DEMOTE.

The idea is good but the implementation needed a few improvements
to make it more reliable. The patch does the following:

- Adds checks that the lock is kept to all the state-changing
  code.

- Takes the lock in some new places where the state was changed
  before without a lock.

- Rename the lock to 'state_latch'. It is active not only during
  promotion, so it can't be named 'promote_latch'.

This will increase the safety of the next commits in scope
of #8095.

NO_DOC=refactoring
NO_CHANGELOG=refactoring
NO_TEST=tested