box: do not apply rebulid-failed grants
Before this patch it was possible to have the role grant make effect
even if the rebuild of effective grants had failed. That was possible
cause setting the user and role maps in `role_grant` and `role_revoke`
wasn't reverted on an error.

The changes can be reverted by the rollback triggers, so let's simply
have them installed before we do the `grant_or_revoke` call.

Closes tarantool/security#152

NO_DOC=security
NO_CHANGELOG=security