say: escape characters when syslog is used in JSON format
This commit makes logger escape prohibited characters in JSON format,
to make Tarantool's syslog logging comply with RFC 3164, which states:

> The MSG part of the syslog packet MUST contain visible (printing)
  characters. The code set traditionally and most often used has also
  been seven-bit ASCII in an eight-bit field like that used in the PRI
  and HEADER parts. In this code set, the only allowable characters
  are the ABNF VCHAR values (%d33 - 126) and spaces (SP value %d32) 1.

Simple `json_escape_inplace` in the end cannot be used for json
escaping, as it was done for plain format, since it'll corrupt the
structure of the message and it won't be possible to decode it after
that.

Closes #12006

@TarantoolBot document
Title: Document the RFC 3164 syslog format compliance
Since: 3.6.1

Documents:
  * https://www.tarantool.io/en/doc/latest/reference/configuration/configuration_reference/#audit-log-syslog
  * https://www.tarantool.io/en/doc/latest/admin/logs/#log-destination
  * https://www.tarantool.io/en/doc/latest/reference/reference_lua/log/
  * https://www.tarantool.io/en/doc/latest/reference/configuration/#confval-log

The syslog format is compliant with
[RFC 3164](https://www.rfc-editor.org/rfc/rfc3164.html). However, the
length of the message can be up to 16 * 1024 bytes.