limbo: associate it with Raft explicitly
The limbo initially was intended as something that can be used
not only as a singleton. Even if never actually needed, this still
enforces some useful practices and design decisions. The best
one - all the related states and configs being visible in one
place as explicit dependencies.

But the limbo was using the global Raft - box_raft(). Which made
it non-obvious of the limbo depends on Raft at all in any way.

Lets store Raft pointer directly in the limbo to make this
relation visible. And this will be quite handy in the next
commits.

In scope of #8095

NO_DOC=refactoring
NO_CHANGELOG=refactoring
NO_TEST=tested

(cherry picked from commit f3f83d385346ffb7b74ff2a0715f02c5d36d59c4)