box: allow to access users if entity rights exist
We have a rule that if an entity stored in a space is readable,
writable, alterable or droppable by a user, then the space contents
are fully visible for him. E. g. if a user has "read" access on the
"space" entity, then it will have access to all spaces in the sysview.

The only exception is the "_vuser" space: we lacked an entity check
there cause these were introduced in the commit af35de967b084d673195
("Introduce privileges for object groups"), but only for spaces,
functions and sequences. So when we've introduced user and role entity
accesses in the commit c46702ff499fd50fc ("Add entities user, role to
access control") the '_vuser' space filter had not been updated.

Let's fix that to have it consistent with other system view spaces.

Closes #12089

NO_DOC=bugfix