box: move collect_confirmed_vclock to replication
This commit moves the `box_collect_confirmed_vclock` to replication.cc
and renames it to `replicaset_collect_confirmed_vclock`.

The function collects remote vclocks, computes the maximum of them in
order to return vclock, that is confirmed by the synchro quorum of a
replicaset. It is more related to the replication, rather than box.

This is done, because in the upcoming commits, where
`replication_linearizable_quorum` will be introduced, the function will
depend on fiber condition, defined in the replication.cc, in order to
allow changing the linearizable quorum on the fly.

Needed for #11876

NO_TEST=refactoring
NO_CHANGELOG=refactoring
NO_DOC=refactoring