box: introduce replication_linearizable_quorum
The option is intended to change the quorum of linearizable transactions
to increase or decrease the factor of quorum reads.

Consider, we have 2 replicas in each DC, 4 replicas in total (N = 4, Q =
3, R = 2). Let's say we have DCs A and B, and connectivity between A and
B is lost. This means Q = 3 isn't reachable anymore, and the cluster
cannot serve writes. In order to restore availability we rely on an
external coordinator to promote the other DC (A) and lower the quorum
(Q) to 2. As soon as this happens, any client, which manages to connect
to a replica from DC B, will observe a non-linearizable read (R = 2 and
the promoted master from DC A has Q = 2, so R and Q do not overlap).

In order to avoid this situation, we want to be able to set R = 3 in
advance. This will be excessive as long as DCs stay interconnected and Q
= 3, but it'll save us from non-linearizable reads in case of DC
connection failure and manual quorum decrease. The replicas from DC B
won't be able to serve linearizable reads as long as they don't see the
replicas from DC A, which's better than returning stale data.

The option is intentionally not explicitly set in the `box_cfg_xc()`
in order to avoid duplicate "update replication_linearizable_quorum"
messages. It's set by the `box_set_replication_synchro_quorum()`.

Closes #11876

@TarantoolBot document
Title: new `replication.linearizable_quorum` option

A number of replicas to synchronize with before performing linearizable
transaction.

Linearizable transactions guarantee that if a write operation completes
before a read operation begins, then this read will always observe the
effects of that write. To ensure this, the node executing a linearizable
transaction must obtain information from enough remote peers to
determine the most up-to-date state of the replica set. That number of
peers can be configured with this option.

By default, the linearizable quorum is evaluated dynamically as
`N - Q + 1`, where: `N` – the current number of registered replicas in
the replica set, `Q` – the current value of `replication.synchro_quorum`.

This default guarantees quorum intersection between synchronous writes and
linearizable reads (`R + Q = N + 1`), which is a fundamental requirement
for linearizability.

Note:
  To begin a linearizable transaction, the MVCC must be enabled and
  `txn_isolation` must be set to `linearizable`, the transaction can only
  perform requests to synchronous, local or temporary spaces.