box: allow to change the replication_linearizable_quorum after txn start
This commit makes changing of replication_linearizable_quorum possible.
Before this commit, the quorum could be changed only when transaction is
yielding in `relay_trigger_vclock_sync` in the `collect_confirmed_vclock`.
Now it's also possible to change the quorum, when waiting for ACKs from
replicas.

However, when vclock is figured out, changing the quorum won't affect
the transaction: we suppose, that if we managed to get ACKs from
vclocks, then it'll be possible to synchronize with these replicas.

Follow-up #11876

NO_DOC=bugfix
NO_CHANGELOG=not released