box: introduce array/table form of field compression
Effectively introduces the `struct field_compression_def` and
`struct field_compression` instead of `enum compression_type`.

The first one defines the compression to be used on the field along
with compression options identical to the ones we have in the Lua
(we have none in CE, but some amount defined in EE).

The second one contains the compression definition along with new
`struct compression_opts` that is meant to be passed to the core
compression routines to perform the compression with the parameters
specified. This is done so that the core library does not depend on
the Tarantool field compression definitions and uses its own API.

Compression value examples:
- `compression = 'lz4'`
- `compression = {'lz4'}`
- `compression = {'lz4', acceleration = 1000}`
- `compression = {'zlib', level = 6, mem_level = 8}`

Closes #12117

NO_DOC=ee
NO_TEST=ee
NO_CHANGELOG=ee