limbo: split "txn queue" and "state"
The limbo was initially intended to be just a dumb simple
transaction queue being fully controlled by an external state
machine of a Raft object.

This by now creates too much problems. Because the limbo has its
own state which is updated even when Raft is not doing anything.
Like when elections are won and persisted in Raft, but the limbo
still needs to catch up eventually and asynchronously and write
its own PROMOTE entry.

On top of that the limbo has lots of other parts of its state
which are intricately moving and entangling with each other like
the fencing, the freeze, the active rollback state, persisted
terms tracking, an ongoing promotion state (svp_confirmed_lsn),
validation being enabled/disabled.

So there are actually 2 logical parts in the limbo: its state
machine and its transaction queue.

The latter is over 50% of the limbo code. About a thousand lines
of code. And it has nothing to do with the state changes. The
queue inside the limbo is really just a dumb list of transactions
with some non-trivial API on top of it.

This creates a lot of confusion during development when there is
almost 2k lines of code in the C file, and one has to go through
it and find where to put queue logic and where to safely make
state changes.

The idea is to split the logic once again. At least until further
refactoring. Factor the queue out of the limbo. So the logic of
the queue and the logic of state changes are separated entirely
and are easier to understand and modify.

The new layout will be very handy in scope of #8095 where the
limbo state will get new complications.

NO_DOC=refactoring
NO_CHANGELOG=refactoring
NO_TEST=tested

(cherry picked from commit 6d825b5d147c24d65e94e49143fed521eddc26be)