limbo: split wait-complete into impl and read-only
The txn_limbo_wait_complete() function does actually 2 things:
the waiting and the rollback. The latter is a legacy bug which
needs to be kept at least until 4.x.

For the purpose of the next commits it will be convenient to split
this function into 2 parts doing these things separately, the
waiting and the rollback.

NO_DOC=refactoring
NO_CHANGELOG=refactoring
NO_TEST=tested