limbo: check fullness only once on confirm
It was previously checked on each confirmed transaction, whether
need to notify any fibers waiting on free space in the limbo.

But can as well do that once in the end of CONFIRM applying.
Anyway the cond is broadcast, not signaled.

NO_DOC=refactoring
NO_CHANGELOG=refactoring
NO_TEST=tested