box: fix crash with transactional trigger on system space
Currently transactional trigger on space `_space` causes crash on DDL
which changes record for `_space` itself in `_space`. In particular this
happens on recovery. Similarly there is a crash on trigger on space
`_index` and changing record for `_index` itself in `_index`.

The issue is we first commit the DDL which destroys the old space and
then run transactional trigger which refers to the old space. Let's do
the opposite, first run transactional trigger.

Closes #11766

NO_DOC=bugfix