memtx: refactor multikey and functional index `replace`
Currently, multikey and functional index `replace` are implemented using an
overload of the `replace` virtual method. But actually this is a crutch,
nothing stops us from implementing them through the regular index replace.
Doing so will improve code health, because it will simplify memtx index
code, reuse common code paths and encapsulate the multikeyness of indexes a
a higher level of abstraction. This, in turn, will further on simplify the
support of multikey indexes in MVCC.

To achieve this, we encapsulate the multikey and functional index
`replace`s as separate paths in `memtx_index_replace`. Internally, we
change the interface of `replace` to use a new abstraction,
`memtx_index_key`, that generalizes the tuple abstraction for multikey and
functional indexes by storing the tuple together with its multikey array
index or functional key to uniquely identify such keys.

While the result key is already used entirely by functional indexes to
rollback operations, the successor key is not, but it will further on allow
us to support multikey indexes in MVCC.

Closes #12190
Needed for #6385

NO_CHANGELOG=<refactoring>
NO_DOC=<refactoring>
NO_TEST=<refactoring>