memtx: reduce false-positive aborts in Memtx-MVCC
It was discovered that Memtx MVCC aborts transactions too frequently, even
in cases where it would seem possible to avoid doing so. The issue was in
how potential duplicates in secondary indexes were handled when a statement
became "prepared".

The processing worked as follows. When a replace/insert statement `stmt`
becomes "prepared" (`stmt->add_story != NULL`):

- Iterate through each chain (except the primary one) upwards from
`stmt->add_story`
- For each `story` along the path, check:
    - If the `story` was added by the same transaction as
      `stmt->add_story`, skip it. The check has already been performed
      for such stories and nothing could have been corrupted.
    - If the `story` was added by a transaction that previously deleted
      (**using DELETE statement**) a tuple with the same key, skip it. It
      couldn't have been a duplicate of `stmt->add_story->tuple` because
      at the serialization point when `story->tuple` was inserted,
      `stmt->add_story->tuple` was definitely deleted.
    - If `story->tuple` has a primary key different from the primary key
      of `stmt->add_story->tuple`
      (`story->add_stmt->del_story != stmt->add_story`), then abort
      `story->add_stmt->txn` - it's a duplicate.

This patch addresses the second point. Actually, it doesn't matter to us
which type of statement deleted a tuple with the same key: a `DELETE` or
a `REPLACE`. Up until now, we have only handled `DELETE`, because its
`del_story` reference is constant and always points to the same tuple,
and the read tracker "monitors" the lifetime of the tuple pointed to by
this reference. Because the reference is set only once, it was easy for
us to handle this case. With `REPLACE` statements, things are a bit more
complicated, as their `del_story` reference can change any number of
times. At different points in time, this property ("the story was added by
a transaction that previously replaced a tuple with the same key") can
toggle between true and false as other transactions become "prepared".
Therefore, we need some mechanism to quickly handle the queries
"transaction `TXN` has become prepared" and "does our property hold for
story `story`?".

Closes #11927, #11942

NO_DOC=enhancement