box: split box_eval_and_check_replication_quorum function
This commit moves the quorum evaluation to the separate function, since
in the next commit it'll be used during update of the linearizable
quorum (e.g. due to the changing the synchro quorum).

This is needed, because there we will want to get the actual quorum,
which was evaluated from a formula and not just error.

Needed for #11876

NO_TEST=refactoring
NO_CHANGELOG=refactoring
NO_DOC=refactoring